+++ src/db_admin_commands.c
@@ -92,7 +92,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
                                    "unlocked_races",   "unlocked_classes",
                                    "player_save_objs", "player_save_objs_sheathed",
                                    "pet_save_objs"};
-    for (i = 0; i < sizeof(player_tables) / sizeof(player_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(player_tables) / sizeof(player_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", player_tables[i]);
     }
@@ -101,7 +101,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
   {
     const char *object_tables[] = {"object_database_items", "object_database_wear_slots",
                                    "object_database_bonuses"};
-    for (i = 0; i < sizeof(object_tables) / sizeof(object_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(object_tables) / sizeof(object_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", object_tables[i]);
     }
@@ -124,7 +124,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
                                        "material_categories",
                                        "material_subtypes",
                                        "material_qualities"};
-    for (i = 0; i < sizeof(wilderness_tables) / sizeof(wilderness_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(wilderness_tables) / sizeof(wilderness_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", wilderness_tables[i]);
     }
@@ -140,7 +140,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
   else if (!str_cmp(system, "ai"))
   {
     const char *ai_tables[] = {"ai_config", "ai_requests", "ai_cache", "ai_npc_personalities"};
-    for (i = 0; i < sizeof(ai_tables) / sizeof(ai_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(ai_tables) / sizeof(ai_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", ai_tables[i]);
     }
@@ -149,7 +149,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
   {
     const char *help_tables[] = {"help_entries", "help_keywords", "help_versions",
                                  "help_search_history", "help_related_topics"};
-    for (i = 0; i < sizeof(help_tables) / sizeof(help_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(help_tables) / sizeof(help_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", help_tables[i]);
     }
@@ -157,7 +157,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
   else if (!str_cmp(system, "region"))
   {
     const char *region_tables[] = {"region_data", "path_data", "region_index", "path_index"};
-    for (i = 0; i < sizeof(region_tables) / sizeof(region_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(region_tables) / sizeof(region_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", region_tables[i]);
     }
@@ -166,7 +166,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
   {
     const char *vessel_tables[] = {"ship_interiors", "ship_docking", "ship_room_templates",
                                    "ship_cargo_manifest", "ship_crew_roster"};
-    for (i = 0; i < sizeof(vessel_tables) / sizeof(vessel_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(vessel_tables) / sizeof(vessel_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", vessel_tables[i]);
     }
@@ -182,7 +182,7 @@ static void show_detailed_database_info(struct char_data *ch, char *system)
                                    "pubsub_message_metadata_v3",
                                    "pubsub_message_fields_v3",
                                    "pubsub_message_tags_v3"};
-    for (i = 0; i < sizeof(pubsub_tables) / sizeof(pubsub_tables[0]); i++)
+    for (i = 0; (size_t)i < sizeof(pubsub_tables) / sizeof(pubsub_tables[0]); i++)
     {
       send_to_char(ch, "  - %s\r\n", pubsub_tables[i]);
     }
